refactoring
- move reduce into the stage for better reusebility
