+++ include/picongpu/fields/poissonSolver/RightHandSideNormalization.hpp
@@ -23,36 +23,13 @@
 #include "picongpu/fields/FieldTmpOperations.hpp"
 #include "picongpu/fields/poissonSolver/FieldV.hpp"
 
-#include <pmacc/device/Reduce.hpp>
 #include <pmacc/lockstep/lockstep.hpp>
 #include <pmacc/mappings/kernel/ExchangeMapping.hpp>
 #include <pmacc/math/operation.hpp>
-#include <pmacc/memory/boxes/DataBoxDim1Access.hpp>
-#include <pmacc/memory/boxes/DataBoxUnaryTransform.hpp>
 #include <pmacc/memory/dataTypes/Mask.hpp>
-#include <pmacc/mpi/MPIReduce.hpp>
-#include <pmacc/mpi/reduceMethods/Reduce.hpp>
 
 namespace picongpu::fields::poissonSolver
 {
-    struct SolutionFunction
-    {
-        HDINLINE auto operator()(math::Vector<double, simDim> const& totalCellCoordinate) const
-        {
-            if constexpr(simDim == 3u)
-            {
-                return math::sin(totalCellCoordinate.x()) + math::cos(totalCellCoordinate.y())
-                       + 3.0 * math::sin(totalCellCoordinate.z())
-                       + totalCellCoordinate.x() * totalCellCoordinate.productOfComponents() + 10.0;
-            }
-            else if constexpr(simDim == 2u)
-            {
-                return math::sin(totalCellCoordinate.x()) + math::cos(totalCellCoordinate.y())
-                       + totalCellCoordinate.x() * totalCellCoordinate.productOfComponents() + 10.0;
-            }
-        }
-    };
-
     struct RightHandSideNormalizationKernel
     {
         DINLINE auto operator()(auto const& worker, auto const fieldVBox, auto fieldRhoBox, auto const mapper) const
@@ -96,14 +73,6 @@ namespace picongpu::fields::poissonSolver
 
     struct RightHandSideNormalization
     {
-        mpi::MPIReduce mpiReduce;
-        std::unique_ptr<pmacc::device::Reduce> localReduce;
-
-        RightHandSideNormalization() : localReduce{std::make_unique<pmacc::device::Reduce>(1024)}
-        {
-            mpiReduce.participate(true);
-        }
-
         // return residual
         // return number of iterations
         void operator()(FieldV& fieldV, FieldTmp& fieldRho, MappingDesc cellDescription)
@@ -123,59 +92,5 @@ namespace picongpu::fields::poissonSolver
                 }
             }
         }
-
-        auto calcNorm(FieldTmp& fieldRho)
-        {
-            /*define stacked DataBox's for reduce algorithm*/
-            using TransformedBox = DataBoxUnaryTransform<typename FieldTmp::DataBoxType, squareComponentWise>;
-            using Box64bit = DataBoxUnaryTransform<TransformedBox, cast64Bit>;
-            using D1Box = DataBoxDim1Access<Box64bit>;
-
-            /* reduce field E*/
-            DataSpace<simDim> fieldSize = fieldRho.getGridLayout().sizeWithoutGuardND();
-            DataSpace<simDim> fieldGuard = fieldRho.getGridLayout().guardSizeND();
-
-            TransformedBox fieldTransform(fieldRho.getDeviceDataBox().shift(fieldGuard));
-            Box64bit field64bit(fieldTransform);
-            D1Box d1Access(field64bit, fieldSize);
-
-            float_64 fieldRhoNormSquaredLocal
-                = (*localReduce)(pmacc::math::operation::Add(), d1Access, fieldSize.productOfComponents()).x();
-
-            // avoid deadlock between not finished pmacc tasks and mpi blocking collectives
-            eventSystem::getTransactionEvent().waitForFinished();
-            float_64 fieldRhoNormSquaredGlobal;
-            mpiReduce(
-                pmacc::math::operation::Add(),
-                &fieldRhoNormSquaredGlobal,
-                &fieldRhoNormSquaredLocal,
-                1,
-                mpi::reduceMethods::AllReduce());
-
-            return math::sqrt(fieldRhoNormSquaredGlobal);
-        }
-
-    private:
-        template<typename T_Type>
-        struct cast64Bit
-        {
-            using result = typename TypeCast<float_64, T_Type>::result;
-
-            HDINLINE result operator()(T_Type const& value) const
-            {
-                return precisionCast<float_64>(value);
-            }
-        };
-
-        template<typename T_Type>
-        struct squareComponentWise
-        {
-            using result = T_Type;
-
-            HDINLINE result operator()(T_Type const& value) const
-            {
-                return value * value;
-            }
-        };
     };
 } // namespace picongpu::fields::poissonSolver
